requests: Split up requests.rs
To facilitate platform abstraction, split requests.rs into:

- kernel/src/sev/caa.rs: implementation for the SVSM Calling Area
- kernel/src/platform/snp_requests.rs: SEV-specific request processing

No functional change intended.
