+++ kernel/src/lib.rs
@@ -30,7 +30,6 @@ pub mod locking;
 pub mod mm;
 pub mod platform;
 pub mod protocols;
-pub mod requests;
 pub mod serial;
 pub mod sev;
 pub mod string;
